requirements.txt
# Web 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic-settings>=2.0.0

# 人脸识别核心库
face-recognition==1.3.0
opencv-python-headless==4.9.0.80
numpy==1.24.3

# 图像处理
Pillow==10.2.0

# 数据库和存储
supabase==2.3.0

# dlib 依赖 (尝试使用预编译版本，或由 face-recognition 自动处理)
dlib==19.24.2
